Added Barcode Search Added Inventory Product adding (not sending to DB yet) fixed major BUGS Cleand-up Classes

+++ ClassLibrary/Objects/product.cs
@@ -10,8 +10,8 @@ namespace ClassLibrary.Objects
     {
         public Guid Id { get; set; }
         public Guid ProductId { get; set; }
-        public float Amount { get; set; }
-        public float Weight { get; set; }
+        public double Amount { get; set; }
+        public double Weight { get; set; }
         public DateTime EP {  get; set; }
     }
 }
